repositories do sistema admin

+++ src/main/java/com/lucas/confeitaria_api/product/entities/RecipeUsage.java
@@ -26,5 +26,5 @@ public class RecipeUsage {
     @JoinColumn(name = "recipe_id")
     private Recipe recipe;
 
-    private BigDecimal quantityUsed;
+    private Double quantityUsed;
 }
